base

+++ addp-web/addp-web-jpa/src/main/java/com/nix/jingxun/addp/web/jpa/MemberJpa.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author keray
  * @date 2019/04/21 0:38
  */
-public interface MemberJpa extends JpaRepository<MemberModel,Integer> {
+public interface MemberJpa extends JpaRepository<MemberModel,Long> {
 }
